+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - All 3 secret chao eggs are now location checks. You have to bring them to the chao garden to get the check.
+- Chao stats multiplier in the mod's settings (default is 1, up to 50).
 
 ### Fixed